parse: call command thunks, and provide way commands to abort
The 'parse' procedure now calls all of the command thunks after parsing.
To accomodate the likes of special-commands, the procedure 'abort-parse'
is provided that can terminate parsing and print the message about
commands that were not called.  Parse returns #t on successful completion,
or #f if parsing was aborted.

With this change, using icla is as simple as this:

  (when (icla:parse (command-line-arguments))
    (start-my-program))

The callinfo accessors are no longer exported.